chore(core): user enabled and endpoint permission checks on every query (#4059)

+++ core/src/main/java/io/questdb/cairo/security/AllowAllSecurityContextFactory.java
@@ -33,7 +33,7 @@ public class AllowAllSecurityContextFactory implements SecurityContextFactory {
     }
 
     @Override
-    public SecurityContext getInstance(CharSequence principal, byte authType, int interfaceId) {
+    public SecurityContext getInstance(CharSequence principal, byte authType, byte interfaceId) {
         return AllowAllSecurityContext.INSTANCE;
     }
 }
